backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.4
email-validator==2.1.1
pyjwt==2.8.0
bcrypt==4.1.3
cryptography==42.0.8
requests==2.31.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
dlib==19.24.6
face_recognition==1.3.0
websockets==12.0
firebase-admin==6.5.0
google-generativeai==0.5.4
